+++ mtalg/core/threads.py
@@ -1,1 +1,2 @@
-_global_num_threads = None
+from multiprocessing import cpu_count
+_global_num_threads = cpu_count()